chatapi/requirements.txt
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
asgiref==3.7.2
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
build==1.0.3
cachetools==5.3.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
coloredlogs==15.0.1
cycler==0.12.1
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
Deprecated==1.2.14
Django==4.0
effdet==0.4.1
emoji==2.10.1
fastapi==0.109.2
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
fsspec==2024.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.1
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.20.3
humanfriendly==10.0
importlib-metadata==6.11.0
importlib-resources==6.1.1
iopath==0.1.10
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.1.5
langchain-community==0.0.17
langchain-core==0.1.18
langchain-openai==0.0.5
langdetect==1.0.9
layoutparser==0.3.4
lxml==4.9.4
MarkupSafe==2.1.5
marshmallow==3.20.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
mypy-extensions==1.0.0
nltk==3.8.1
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
opencv-python==4.9.0.80
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
overrides==7.7.0
packaging==23.2
pdf2image==1.17.0
pikepdf==8.12.0
pillow==10.2.0
portalocker==2.8.2
pulsar-client==3.4.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pyparsing==3.1.1
pypdf==4.0.1
PyPika==0.48.9
pyproject_hooks==1.0.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==0.21.1
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.6.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
starlette==0.36.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.2
timm==0.9.12
tomli==2.0.1
torch==2.2.0
torchvision==0.17.0
transformers==4.37.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
unstructured==0.11.8
unstructured-client==0.17.0
unstructured-inference==0.7.23
uvloop==0.19.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
zipp==3.17.0
